// src/generated/newspaper.schemas.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Swagger Petstore - OpenAPI 3.0
 * OpenAPI spec version: 1.0.11
 */
export interface Box {
    children?: ContentType
}

export interface Section {
    text?: string
}

export interface Paragraph {
    text?: string
}

export type ContentType = Paragraph | Section | Box

export interface Article {
    author?: string
    content?: ContentType
    id?: number
    title?: string
    topic?: string
}

export interface ArticleResponse {
    articles?: Article[]
}